# api/openapi.yaml
openapi: 3.0.3
info:
  title: OwnDock API
  version: 0.1.0
  description: Pre-release browser, management, and automation HTTP contract. Identity, the read-only built-in Template catalog, Managed Host and Agent enrollment, Project-scoped Application, Release, Registry Credential, Source Repository connection, Build Configuration, manual, token, and signed Webhook Build trigger/query, bounded incremental Build logs, Artifact, read-only digest-bound Artifact Evidence metadata and verified content download, Release handoff, Environment, Runtime Target, Deployment lifecycle, Runtime Inventory, controlled Terminal Session, and Audit operations form the implemented product slice when product mode is enabled. Every formal Product API operation is subject to shared source and installation-wide ingress admission and can return 429 rate_limited or fail-closed 503 ingress_protection_unavailable with Retry-After. Browser access is same-origin by default; configured cross-origin clients use exact origins without REST authentication cookies, and every `/api/` response is non-cacheable. Terminal connection tickets are short-lived, one-time, path-scoped Secure HttpOnly cookies and are never returned in a URL or JSON body. The separate mTLS Agent Control v1 transport is documented outside this OpenAPI; its typed commands and Agent Runtime Gateway are internal transports. Top-level Application, Environment, and Deployment operations remain disabled engineering samples.
...
servers:
  - url: /
...
paths:
...
  /api/v1/managed-hosts:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Managed Hosts]
      operationId: listManagedHosts
      summary: List Organization-owned managed hosts
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ManagedHostList"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Managed Hosts]
      operationId: createManagedHost
      summary: Register an Organization-owned managed host
      description: Only an Owner may register a host. Agent hosts remain enrolling until the separate enrollment flow creates a fixed Agent identity, then remain offline until the authenticated Agent Control flow establishes liveness.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManagedHostRequest"
      responses:
        "201":
          $ref: "#/components/responses/ManagedHostCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
  /api/v1/managed-hosts/{managed_host_id}:
    parameters:
      - $ref: "#/components/parameters/RequestID"
      - $ref: "#/components/parameters/ManagedHostID"
    get:
      tags: [Managed Hosts]
      operationId: getManagedHost
      summary: Get an Organization-owned managed host
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ManagedHostResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/terminal-policy:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Terminal Sessions]
      operationId: getOrganizationTerminalPolicy
      summary: Get the Organization host-terminal policy
      description: Owner only. When no policy has been saved, returns the restrictive default that allows host terminals only for Owners.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/TerminalPolicyResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
    put:
      tags: [Terminal Sessions]
      operationId: updateOrganizationTerminalPolicy
      summary: Create or replace the Organization host-terminal policy
      description: Owner only. expected_version is 0 when saving the returned default for the first time, then must match the persisted version.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalPolicyRequest"
      responses:
        "200":
          $ref: "#/components/responses/TerminalPolicyResponse"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/terminal-sessions/{terminal_session_id}:
    parameters:
      - $ref: "#/components/parameters/RequestID"
      - $ref: "#/components/parameters/TerminalSessionID"
    get:
      tags: [Terminal Sessions]
      operationId: getTerminalSession
      summary: Get safe Terminal Session metadata
      description: Returns lifecycle and fixed target references, never the one-time ticket, its hash, terminal input, or terminal output.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/TerminalSessionResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/agent/enrollments:exchange:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    post:
      tags: [Managed Hosts]
      operationId: exchangeAgentEnrollment
      summary: Exchange a one-time enrollment token and CSR for an Agent certificate
      description: This bootstrap endpoint does not accept a user session. The enrollment token is carried only in the JSON body and is consumed atomically. For response-loss recovery, the exact same token, CSR, instance, version, protocol and capability request may retrieve the original identity and certificate for 10 minutes; any changed request is rejected and no second identity or audit event is created.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeAgentEnrollmentRequest"
      responses:
        "201":
          $ref: "#/components/responses/AgentCredentialsCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "503":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/projects:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects in the current organization
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProjectList"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNamedResourceRequest"
      responses:
        "201":
          $ref: "#/components/responses/ProjectCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
  /api/v1/templates:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Templates]
      operationId: listBuiltInTemplates
      summary: List immutable built-in Application presets
      description: Returns the community catalog. Templates carry no credentials or target bindings and cannot mutate existing Applications.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/TemplateList"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
  /api/v1/templates/{template_id}:
    parameters:
      - $ref: "#/components/parameters/RequestID"
      - $ref: "#/components/parameters/TemplateID"
    get:
      tags: [Templates]
      operationId: getBuiltInTemplate
      summary: Get one immutable built-in Application preset
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/TemplateItem"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/build-triggers/{build_trigger_id}:
    parameters:
      - $ref: "#/components/parameters/RequestID"
      - $ref: "#/components/parameters/BuildTriggerID"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Build Triggers]
      operationId: triggerExternalBuild
      summary: Enqueue a Build from any Git hosting automation
      description: Send the one-time Build Trigger token as Authorization Bearer. The caller can supply only the exact ref and full commit SHA; repository, Build Configuration, registry, and allowed refs come from the stored trigger. The Server verifies that the ref currently resolves to the supplied commit before accepting the Build. The shared default limit is 60 authenticated requests per trigger per minute.
      security:
        - buildTriggerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerExternalBuildRequest"
      responses:
        "202":
          $ref: "#/components/responses/BuildAccepted"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "429":
          $ref: "#/components/responses/BuildTriggerRateLimited"
        "503":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
...
  /api/v1/audit-events:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: List audit events for the current organization
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        "200":
          $ref: "#/components/responses/AuditEventList"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
  /api/v1/applications:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Applications]
      operationId: listApplications
      summary: List applications
      responses:
        "200":
          $ref: "#/components/responses/ApplicationList"
        "500":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Applications]
      operationId: createApplication
      summary: Create an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          $ref: "#/components/responses/ApplicationCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
  /api/v1/environments:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Environments]
      operationId: listEnvironments
      summary: List environments
      responses:
        "200":
          $ref: "#/components/responses/EnvironmentList"
        "500":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Environments]
      operationId: createEnvironment
      summary: Create an environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentRequest"
      responses:
        "201":
          $ref: "#/components/responses/EnvironmentCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
  /api/v1/deployments:
    parameters:
      - $ref: "#/components/parameters/RequestID"
    get:
      tags: [Deployments]
      operationId: listDeployments
      summary: List deployments
      parameters:
        - name: application_id
          in: query
          required: false
          schema:
            type: string
        - name: environment_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/DeploymentList"
        "500":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Deployments]
      operationId: createDeployment
      summary: Create a deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentRequest"
      responses:
        "201":
          $ref: "#/components/responses/DeploymentCreated"
        "400":
          $ref: "#/components/responses/APIError"
        "415":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "500":
          $ref: "#/components/responses/APIError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque user session token sent explicitly in Authorization. The REST API does not authenticate user cookies.
    terminalTicket:
      type: apiKey
      in: cookie
      name: __Secure-owndock_terminal_ticket
      description: Short-lived, one-time terminal ticket scoped to the exact connect path. It is never returned in JSON or accepted in the URL.
    buildTriggerAuth:
      type: http
      scheme: bearer
      bearerFormat: OwnDock-Build-Trigger
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller-provided request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._-]+$"
...
    ManagedHostID:
      name: managed_host_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    TerminalSessionID:
      name: terminal_session_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
...
    ManagedHostList:
      description: Organization managed host collection.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [items]
            properties:
              items:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedHost"
    ManagedHostCreated:
      description: Managed host registered.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManagedHost"
    ManagedHostResponse:
      description: Managed host state.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ManagedHost"
...
    AgentCredentialsCreated:
      description: Newly issued Agent client identity and certificate chain.
      headers:
        Cache-Control:
          schema:
            type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentCredentials"
...
    APIError:
      description: Request failed. The safe message is localized from Accept-Language; clients branch on the stable error.code, never on message text.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Content-Language:
          description: Language of the safe error message selected from Accept-Language. Unsupported or missing languages fall back to en-US.
          schema:
            type: string
            enum: [en-US, zh-CN]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TerminalPolicyResponse:
      description: Effective Terminal access policy. A version of 0 means the restrictive default has not been persisted.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TerminalPolicy"
    TerminalSessionResponse:
      description: Safe Terminal Session metadata with no connection ticket or terminal payload.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TerminalSession"
...
  schemas:
...
    TerminalPolicyRequest:
      type: object
      additionalProperties: false
      required: [enabled, allowed_roles, environment_stages, runtime_target_ids, managed_host_ids, idle_timeout, maximum_duration, maximum_per_user, maximum_per_target, revocation_grace_period, expected_version]
      properties:
        enabled:
          type: boolean
        allowed_roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [owner, maintainer, developer]
        environment_stages:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [development, staging, production]
        runtime_target_ids:
          type: array
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 160
        managed_host_ids:
          type: array
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 160
        idle_timeout:
          type: string
          example: 10m
        maximum_duration:
          type: string
          example: 1h
        maximum_per_user:
          type: integer
          minimum: 1
          maximum: 10
        maximum_per_target:
          type: integer
          minimum: 1
          maximum: 50
        revocation_grace_period:
          type: string
          example: 30s
        expected_version:
          type: integer
          format: int64
          minimum: 0
...
    ExchangeAgentEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [enrollment_token, instance_id, agent_version, protocol_version, capabilities, csr_pem]
      properties:
        enrollment_token:
          type: string
          minLength: 32
          writeOnly: true
        instance_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9_-][A-Za-z0-9._-]*$"
        agent_version:
          type: string
          minLength: 1
          maxLength: 64
        protocol_version:
          type: string
          minLength: 1
          maxLength: 64
        capabilities:
          description: Capabilities approved for this Agent Identity. A later authenticated hello may advertise only a subset of this grant.
          type: array
          maxItems: 16
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 64
            pattern: "^[a-z0-9][a-z0-9._-]{0,63}$"
        csr_pem:
          type: string
          minLength: 1
          maxLength: 16384
          writeOnly: true
...
    CreateManagedHostRequest:
      oneOf:
        - $ref: "#/components/schemas/CreateAgentManagedHostRequest"
        - $ref: "#/components/schemas/CreateDirectManagedHostRequest"
...